move robot_state_tap.h into systems/plants
