+++ drake/examples/kuka_iiwa_arm/run_kuka_iiwa_arm_dynamics.cc
@@ -1,12 +1,12 @@
 #include <iostream>
 
 #include "drake/examples/kuka_iiwa_arm/iiwa_simulation.h"
-#include "drake/examples/kuka_iiwa_arm/robot_state_tap.h"
 #include "drake/systems/LCMSystem.h"
 #include "drake/systems/Simulation.h"
 #include "drake/systems/cascade_system.h"
 #include "drake/systems/plants/BotVisualizer.h"
 #include "drake/systems/plants/RigidBodySystem.h"
+#include "drake/systems/plants/robot_state_tap.h"
 #include "drake/common/polynomial.h"
 #include "drake/util/drakeAppUtil.h"
