+++ drake/examples/kuka_iiwa_arm/test/kuka_iiwa_arm_test.cc
@@ -6,6 +6,7 @@
 #include "drake/examples/kuka_iiwa_arm/iiwa_simulation.h"
 #include "drake/systems/LCMSystem.h"
 #include "drake/systems/plants/BotVisualizer.h"
+#include "drake/systems/plants/robot_state_tap.h"
 #include "drake/util/eigen_matrix_compare.h"
 
 using drake::RigidBodySystem;
