+++ drake/examples/kuka_iiwa_arm/test/kuka_iiwa_pd_control_test.cc
@@ -4,12 +4,12 @@
 
 #include "drake/common/drake_path.h"
 #include "drake/examples/kuka_iiwa_arm/iiwa_simulation.h"
-#include "drake/examples/kuka_iiwa_arm/robot_state_tap.h"
 #include "drake/systems/LCMSystem.h"
 #include "drake/systems/LinearSystem.h"
 #include "drake/systems/cascade_system.h"
 #include "drake/systems/pd_control_system.h"
 #include "drake/systems/plants/BotVisualizer.h"
+#include "drake/systems/plants/robot_state_tap.h"
 #include "drake/util/eigen_matrix_compare.h"
 
 using drake::AffineSystem;
