+++ drake/systems/plants/robot_state_tap.h
